+++ src/rest/endpoints/webhook.zig
@@ -327,7 +327,7 @@ pub const ExecuteWebhookJsonBody = struct {
     embeds: jconfig.Omittable(?[]const model.Message.Embed) = .omit,
     allowed_mentions: jconfig.Omittable(?model.Message.AllowedMentions) = .omit,
     components: jconfig.Omittable(?[]const model.MessageComponent) = .omit,
-    attachments: jconfig.Omittable(?[]const jconfig.Partial(model.Message.Attachment)) = .omit,
+    attachments: jconfig.Omittable(?[]const PartialAttachment) = .omit,
     flags: jconfig.Omittable(?model.Message.Flags) = .omit,
     thread_name: jconfig.Omittable(?[]const u8) = .omit,
     applied_tags: jconfig.Omittable(?[]const model.Snowflake) = .omit,
@@ -345,7 +345,7 @@ pub const ExecuteWebhookFormBody = struct {
     allowed_mentions: ?model.Message.AllowedMentions = null,
     components: ?[]const model.MessageComponent = null,
     files: ?[]const ?rest.Upload = null,
-    attachments: ?[]const jconfig.Partial(model.Message.Attachment) = null,
+    attachments: ?[]const PartialAttachment = null,
     flags: ?model.Message.Flags = null,
     thread_name: ?[]const u8 = null,
     applied_tags: ?[]const model.Snowflake = null,
@@ -364,7 +364,7 @@ pub const EditWebhookMessageJsonBody = struct {
     flags: jconfig.Omittable(?model.Message.Flags) = .omit,
     allowed_mentions: jconfig.Omittable(?model.Message.AllowedMentions) = .omit,
     components: jconfig.Omittable(?[]const model.MessageComponent) = .omit,
-    attachments: jconfig.Omittable(?[]const jconfig.Partial(model.Message.Attachment)) = .omit,
+    attachments: jconfig.Omittable(?[]const PartialAttachment) = .omit,
     poll: jconfig.Omittable(?model.Poll) = .omit, // Polls can only be added when editing a deferred interaction response.
 
     pub const jsonStringify = jconfig.stringifyWithOmit;
@@ -377,6 +377,14 @@ pub const EditWebhookMessageFormBody = struct {
     allowed_mentions: ?model.Message.AllowedMentions = null,
     components: ?[]const model.MessageComponent = null,
     files: ?[]const ?rest.Upload = null,
-    attachments: ?[]const jconfig.Partial(model.Message.Attachment) = null,
+    attachments: ?[]const PartialAttachment = null,
     poll: ?model.Poll = null, // Polls can only be added when editing a deferred interaction response.
+};
+
+pub const PartialAttachment = struct {
+    id: u64,
+    filename: []const u8,
+    description: jconfig.Omittable([]const u8) = .omit,
+
+    pub const jsonStringify = jconfig.stringifyWithOmit;
 };